+++ contracts/src/models/beast_stats.cairo
@@ -18,8 +18,8 @@ pub struct FixedBeastStats {
 pub struct LiveBeastStats {
     #[key]
     pub token_id: u32,
-    pub current_health: u16,
-    pub bonus_health: u16,
+    pub current_health: u32,
+    pub bonus_health: u32,
     pub last_death_timestamp: u64,
     pub num_deaths: u16,
     pub last_killed_by: u32,
